Alternative version of script to run old dcprogs
Needs autohotkey. Just presses enter 75 times in a row. The problem is
that there is no good way to detect if the program is done with a
calculation before continuing. At the moment it tries to look for
changes in the file size of the output but it is not that good.